feat(room): implement room management CRUD and normalize id types

+++ src/main/java/DAJ2EE/demo/repository/PermissionRepository.java
@@ -4,6 +4,6 @@ import DAJ2EE.demo.entity.Permission;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
-public interface PermissionRepository extends JpaRepository<Permission, Long> {
+public interface PermissionRepository extends JpaRepository<Permission, Integer> {
     Optional<Permission> findByName(String name);
 }
